Mercado_Cuevas_Api/src/pyproject.toml
[tool.poetry]
name = "mercado_cuevas_api"
version = "0.1.0"
description = "mercado cuevas api(delivery app)"
authors = ["aaronI45 <aaron.145@hotmail.com>"]
readme = "../README.md"
packages = [{ include = "mercado_cuevas_api" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
requests = "^2.31.0"
firebase-admin = "^6.2.0"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
poetry = "^1.6.1"
pydantic = "^2.4.2"
json5 = "^0.9.14"
pymongo = "^4.6.0"
python-dotenv = "^1.0.0"
authlib = "^1.2.1"
phonenumbers = "^8.13.26"
pydantic-extra-types = "^2.1.0"
email-validator = "^2.1.0.post1"
pydantic-mongo = "^2.0.3"
google-cloud-storage = "^2.14.0"
python-multipart = "^0.0.6"
fast-api-gcs = "^0.0.11"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
protobuf = "^4.25.1"
passlib = "^1.7.4"
python-jose = { extras = ["criptografia"], version = "^3.3.0" }
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
